requirements.txt
# Core dependencies
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.8.0

# Computer Vision
opencv-python==4.10.0.84
numpy==1.26.4
Pillow==10.4.0

# AI Models
ultralytics==8.3.0  # YOLO
transformers==4.44.0  # For LLaVA
torch==2.4.0
torchvision==0.19.0
segment-anything==1.0  # MobileSAM base

# Voice/Audio
pipecat-ai==0.0.40
vosk==0.3.45
piper-tts==1.2.0
sounddevice==0.5.0
scipy==1.14.0

# Configuration
configparser==7.0.0
python-dotenv==1.0.1

# Analytics
posthog==3.5.0

# Utilities
requests==2.32.3
tqdm==4.66.5
psutil==6.0.0